+++ src/Results.ts
@@ -1,5 +1,5 @@
-import type { Checker } from './Checker'
-import type { Tag } from './Tag'
+import type { Checker } from './Checker.js'
+import type { Tag } from './Tag.js'
 
 class Results {
   append (tag: Tag): void {
@@ -7,10 +7,10 @@ class Results {
   }
 
   public readonly element: HTMLElement
-  constructor (readonly parent: Checker) {
-    const element = document.getElementById('checker.results')
+  constructor (readonly parent: Checker, elementId: string = 'checker.results') {
+    const element = document.getElementById(elementId)
     if (element == null) {
-      throw new Error('Element with Id "checker.results" not found.')
+      throw new Error(`Element with Id "${elementId}" not found.`)
     }
     this.element = element
   }