+++ CHANGELOG.md
@@ -1,3 +1,50 @@
+## [1.17.0](https://github.com/ipfs/public-gateway-checker/compare/v1.16.3...v1.17.0) (2025-11-30)
+
+### Features
+
+* add support for Tor .onion gateways ([#651](https://github.com/ipfs/public-gateway-checker/issues/651)) ([e145c95](https://github.com/ipfs/public-gateway-checker/commit/e145c95381337767c8e4ca84470cf5237d97b29e))
+
+## [1.16.3](https://github.com/ipfs/public-gateway-checker/compare/v1.16.2...v1.16.3) (2025-11-05)
+
+### Bug Fixes
+
+* **ci:** skip ipns and dnslink tests in conformance ([#640](https://github.com/ipfs/public-gateway-checker/issues/640)) ([1e24fbb](https://github.com/ipfs/public-gateway-checker/commit/1e24fbb78e308564fe9a118cda3b50b7d1f9f7bb))
+
+## [1.16.2](https://github.com/ipfs/public-gateway-checker/compare/v1.16.1...v1.16.2) (2025-11-05)
+
+### Trivial Changes
+
+* update ipfs.io and discuss.ipfs.io links to .tech domains ([7e93624](https://github.com/ipfs/public-gateway-checker/commit/7e93624d36372ef9c5d5289504873a5c46ff4372))
+
+## [1.16.1](https://github.com/ipfs/public-gateway-checker/compare/v1.16.0...v1.16.1) (2025-11-04)
+
+### Bug Fixes
+
+* update logo path ([#638](https://github.com/ipfs/public-gateway-checker/issues/638)) ([0b82f20](https://github.com/ipfs/public-gateway-checker/commit/0b82f2003f163f0929184bc80e0bd5cf997626aa))
+
+### Trivial Changes
+
+* **deps-dev:** bump aegir from 42.2.11 to 46.0.0 ([#632](https://github.com/ipfs/public-gateway-checker/issues/632)) ([23fe01a](https://github.com/ipfs/public-gateway-checker/commit/23fe01acb57bca97d17e3c04307763d3fb7036d8))
+
+## [1.16.0](https://github.com/ipfs/public-gateway-checker/compare/v1.15.2...v1.16.0) (2025-11-03)
+
+
+### Features
+
+* add dget.top ([#635](https://github.com/ipfs/public-gateway-checker/issues/635)) ([2ce24b5](https://github.com/ipfs/public-gateway-checker/commit/2ce24b5909e527f4f6c6b91ee97fbcc40f9c99c1))
+
+
+### Trivial Changes
+
+* update conformance results ([92a9763](https://github.com/ipfs/public-gateway-checker/commit/92a9763e2bb5ea736c670ddd37037408aeeee563))
+
+## [1.15.2](https://github.com/ipfs/public-gateway-checker/compare/v1.15.1...v1.15.2) (2025-03-24)
+
+
+### Bug Fixes
+
+* public url ([22b5198](https://github.com/ipfs/public-gateway-checker/commit/22b51985454490b4565ed60427290786975ba0a3))
+
 ## [1.15.1](https://github.com/ipfs/public-gateway-checker/compare/v1.15.0...v1.15.1) (2025-02-19)